RaspberryPi/requirements.txt
matplotlib
numpy
python-can

# for gui
pyqt5
pyqtgraph